Rescaling the time constant in MomentumConn
The MomentumConn parameter momentumTau has been deprecated, in
favor of timeConstantTau. If using timeConstantTau, the
update rule scales the Hebbian term dWMax*pre*post by (1-tauFactor),
where tauFactor is exp(-1/tau) for viscosity and tau otherwise.

The reason for the rescaling is so that, with no momentumDecay,
the asymptotic behavior of the weight in response to a single
pre*post event is to converge to dWMax*pre*post.

If timeConstantTau is present in params, momentumTau is ignored.
If both are absent, defaults for timeConstantTau are used.
If timeConstantTau is absent and momentumTau is present, the old
behavior is still used, but warnings are issued.