Class name changes
PublisherComponent is now BasePublisherComponent.
SparseLayerFlagPublisherComponent (used by almost all layer types) is
now PublisherComponent.

A reminder that the difference between the two classes is that
PublisherComponent (the former SparseLayerFlagPublisherComponent) has
a sparseLayer parameter but BasePublisherComponent does not
(although the data member mSparseLayer flag exists in the
BasePublisherComponent class, it is always false unless a derived
class changes it).