Adds .kpp file to reduce code duplication
The .cu file includes the .kpp file with the preprocessor directive
PV_RUN_ON_GPU set, which creates a function for use by the CUDA kernel.
The .cpp file includes .kpp with the directive unset, which creates
a function for use by the CPU. These two functions share the same
core code.