HyPerDeliveryFacade has been eliminated
Instead, theree is a class HyPerDeliveryCreator.
HyPerConn::createDeliveryObject creates a HyPerDeliveryCreator
object and adds it to the component table.
HyPerDeliveryCreator has a create() method that returns a
HyPerDelivery object (created by the new operator).
HyPerConn calls this method and adds the resulting HyPerDelivery
object to the component table.

HyPerDeliveryCreator has no behavior apart from reading and writing
the params used to decide which HyPerDelivery subclass to create,
and implementing the create() method. This is simpler than
HyPerDeliveryFacade, which had to do all the above, as well as forward
messages from HyPerConn to the HyPerDelivery object. These changes
shorten the code by about 200 lines.