Adds an option to shuffle params groups
Add -shuffle <seed> to the command line, where seed is a positive
integer. This will randomly shuffle the order of the ParameterGroup
objects after PV_Init reads the parameters. Adding -shuffle 0
has no effect. If the -shuffle option is listed, it must be followed
by a nonnegative integer.

Under CMake, you can set the variable PV_SYSTEM_TEST_SHUFFLE to the
desired value to have the system tests run with params files shuffled.
Leave this variable blank to run without the shuffle option.

This option is intended for debugging purposes, to help verify that
the objects do not depend on the order they appear in the params file
(with the exception that HyPerCol has to be first).