Hotfix: Bug in adding probes to ColumnEnergyProbe
If a probe that adds to a column energy probe is postponed during the
communicate stage, it was adding to the total energy each time its
communicateInitInfo() called. The probe would be added each time
the function was called. This fix adds a flag so that the probe
does not add itself multiple times.

12345678901234567890123456789012345678901234567890