PV_CUDA_ARCHITECTURE cmake variable
This removes the hardcoded gpu-architecture setting from the
PVConfigProject.cmake file and provides a CMake variable to
control the setting.

PV_CUDA_ARCHITECTURE is a string that can be set to an NVIDIA compute
capability, such as 7.0, or an NVIDIA microarchitecture name, such
as Volta. If set to Auto, it will detect the compute capability of
the current GPU and use that. If left blank, it will use the
default setting for the installed nvcc compiler's version.

More technically, the PV_CUDA_ARCHITECTURE is passed to the
cuda_select_nvcc_arch_flags() function in CMake's FindCUDA module,
and the result is added to the nvcc options.

Note, however, that FindCUDA is deprecated, and there is not currently
an un-deprecated replacement for cuda_select_nvcc_arch_flags().