recordSize field of .pvp header is no longer used
The recordSize field of the .pvp header is susceptible to arithmetic
overflow for large problems. It can also be calculated as needed
from other fields in the header, for both layer and connection
PVP types.

With this commit, the recordSize value is stored in the header as
it was previously, but code making use of the recordSize field has
been replaced with code that computes it from the rest of the header
information.