Allows negative frame numbers in params files
The frameNumber param used when initializing a layer or connection from
a file can be set to a negative number; in this case the frame index is
counted down from the end of the file.  That is, -1 corresponds to the
last frame, -2 to the second-last, and so on.

This also applies to the initPrev_dWFile parameter for initializing
the previous delta-W values in MomentumConn.

There are two new system tests,
NegativeConnFrameNumber and NegativeLayerFrameNumber, that test
this feature.