Tweaks to CheckpointWeightTest
Test does not use (and has not been using) the files in the input
directory, including the .params files; hence the input directory has
been deleted, and the CMakeLists.txt file specifies NO_PARAMS.
This means only a single CheckpointWeightTest is run for each number
of processors, instead of running both CheckpointWeightTestShared and
CheckpointWeightTestNonshared. The single test evaluates both the
shared and nonshared weights.

If any discrepancies in reading the weights from checkpoint are found,
the difference is printed as well as the expected and observed values,
in order to show differences to small too appear when printf'ed with %f.