Fix to avoid false positive caused by roundoff
CheckpointWeightTest now loads the checkpoint into a separate
Weights object, and compares the result to the original weight.

The previous behavior, reading the weights into the original weights
object, and recalculating the correct value for comparison, was
leading to roundoff issues, even though the calculation being used
was extremely simple.