Probe files remember positions when checkpointing
Files created using the probeOutputFile parameter in probes now
checkpoint their file positions. Thus, starting from a checkpoint
using the -r or -c command line options (or their corresponding
equivalents in a config file) causes the output file to begin
writing at the place it was when the checkpoint was written.