CheckpointableFileStream remembers file positions
To update the file positions every time the file is written,
PrintStream::operator<< and PrintStream::printf were made virtual, so
that CheckpointableFileStream can override them.

A disadvantage is that member function templates cannot be virtual,
so the general operator<< function template cannot be overridden.
Instead, each standard argument type is declared as its own overload.

There is also a change to eliminate a duplicate infomation message
in Checkpointer.