New classes MPIRecvStream and MPISendStream
MPISendStream is a PrintStream-derived class to handle sending
information over MPI using the PrintStream interface.

MPIRecvStream is a class to handle receiving the information sent
by MPISendStream.

The intended use case is for probes to be able to print information
using the standard output streams even if the process that knows the
probe information is not a root process for input/output.

There is also a system test, MPISendStreamTest.