BaseConnectionProbe and PointProbe are deprecated
Tests that use these probes in the BaseConnectionProbe or PointProbe
hierarchy, but are not meant to test those probes specifically, have
been rewritten to use other probes.

The PointProbeTest and PointLIFProbeTest system tests have been moved
from the tests/ directory to the deprecated-tests/ directory.
tests/CMakeLists.txt still builds the deprecated tests, but they may
be removed in a future update.

The UpdateFromCloneTest system test duplicates the coverage of
PlasticCloneConnTest, and appears to be based on a long-removed
HyPerConn parameter. This test has been deleted.