Enhancements and bugfixes for probe output
When run under MPI, only the root process(es) writes probe output to
disk. If a process has row 0, column 0 in the MPI block, but not
batch index 0, that process computes probe output but does not write
the output directly to disk. Instead, it communicates with the
process with batch index 0, using MPISendStream and MPIRecvStream.

Receives are nonblocking, and there is buffering so that a given
line of output does not have to be received on the same timestep that
it was sent. The size of the buffer is currently hardwired at 10.

There is a new system test, ProbeOutputTest. It runs under several
MPI configurations, and checks that the generated probe output files
are the same as the precomputed correct values, allowing for floating
point round-off.