New FileManager class and associated system test
FileManager is designed to have a relatively simple public interface,
and manage the M-to-N directories so that objects that do file I/O
through the FileManager never have to care about how the directories
are managed.

Each MPIBlock's root process works in a separate directory, and the
function members always interpret paths passed to them as relative
to that directory. Importantly, objects using the FileManager do
not need to know this directory if all file operations are handled
using the FileManager, and there is no accessor function to retrieve
that directory. (There is a makeOutputPathFilename() function member
that can be used to retrieve the directory, but it is a transitional
function member, and it is deprecated.)

There is a new system test, FileManagerTest, for the FileManager class.