Deletes parser-generated directory
This directory is generated as needed in the build tree at compile time.
Out-of-source builds are recommended, so this directory is not needed
in the source tree. A .gitignore file has been added to the src/io/
directory, so that an in-source build will not lead to the unneeded
directory being added to the repository.