Adds clobberFlag to MPIRecvStream constructor
If an MPIRecvStream object is instantiated with clobberFlag set to true,
the file will be opened in write-only mode and then closed, thereby
clobbering the file if it had previously existed. It then opens the file
in read/write mode.

If clobberFlag is set to false, MPIRecvStream behaves as before,
creating the file if it does not already exist, and then opening the
file in read/write mode.