Updates interface for timers pie chart creator
python/timerPie.py now takes the input file on the command line
instead of having it hard-coded. The output PNG files are written
to the same directory as the input file, instead of having the
outputDirectory hard-coded as well.

Logic has also been added so that StatsProbes do not cause an error.
The io timer that the StatsProbe inherits from LayerProbe is included
in the output, but the StatsProbe subtimers, whose message starts with
"StatsProbe", are not included.