Updates to HyPerCol
initialize_base() has been removed in favor of initializing data
members in their declarations.

The git revision string is managed by a new class, GitRevision.
This means that when the git status changes but HyPerCol is not
otherwise affected, HyPerCol.cpp does not have to be recompiled;
only GitRevision.cpp does.