Adds two column timers, build and buildrun.
The build timer starts at the beginning of the HyPerCol::run() method,
and stops just before that method calls the advanceTimeLoop method.
Hence it covers the initialization before the first timestep.

The buildrun timer starts at the beginning of the HyPerCol::run()
method as well, and ends just before that method returns. Hence it
covers everything the build timer covers, as well as the
advanceTimeLoop() method.

The build timer, run timer, and checkpoint timer do not overlap, and
together they cover essentially the entire HyPerCol::run() method.
Hence the sum of the times of these three timers should be very close
to the time of the buildrun timer.

The Timer::elapsed_time() method was modified so that it can be used
in the interval after the start() method is called and before the
stop method is called. In this case it functions like the lap button
on a typical stopwatch.