Adds python/timers_to_pie.py
This script creates pie charts of the timers inside the "run" timer.
It is intended to be an enhancement of timerPie.py and to eventually
replace it. It arranges the pie slices by size, creates a slice for
time in the run loop not currently measured by any other timer,
and lumps all slices under a certain threshold into a single slice
labeled "Other".