New InitV and InitWeights classes
DiscreteUniformRandomV is a new InitV class.
InitDiscreteUnifromRandomV is a new InitWeights class.

Each of these classes initializes their object with random values,
where the random values are uniformly distributed from a finite number
of evenly spaced values. The classes have params that specify the
minimum and maximum values of the distribution and the number of
possible values.