Improves GPUSystemTest_postTestOneToMany
GPUSystemTest_postTestOneToMany uses DiscreteUniformRandomV and
InitDiscreteUniformRandomWeights to initialize the input values
and weights with values from the set
   1, 1+1/32, 1+2/32, ..., 1+31/32
This means that the result of the convolution can be represented
exactly[0] in single-precision floating point, removing the need for
an empirically determined tolerance.

The probe in the params file for this test has been replaced with
a RequireAllZeroActivity probe with nnzTolerance of zero.

[0]In the worst case, one needs one more bit than single-precision,
but the result has to be more than 100 standard deviations greater
than the expected value for this to happen.