GPUSystemTest_postTestOneToMany, again
The previous commit changed GPUSystemTest_postTestOneToMany so that
the exact answer could be represented exactly in single-precision
floating point and therefore expected that the test could be
implemented without an allowance for roundoff error.

This overlooked that the GPU computation uses cuDNN library routines,
which do not necessarily perform the convolution using the definition
of a sum of products. There must still be a positive threshold when
comparing the CPU and GPU values.

In this case, the exact values of the convolution have an expected
value of approximately 11400 and therefore the least significant
bit of the single-precision mantissa has value ~= 1e-3. On this
basis, nnzThreshold has been set to 4e-3.