Overhaul of Probes
This commit refactors the probe classes with two aims in mind:
* to implement a more component-based model for managing M-to-N
  communication, and
* to recognize that StatsProbe is different enough from other basic
  probe classes that it complicates matters to try to force it into
  deriving from a common interface, as was done previously.

StatsProbe now derives directly from BaseObject.

ColumnEnergyProbe and the probes for taking norms over layers are
derived from a pure virtual class called ProbeInterface, that provides
the interface for AdaptiveTimeScaleProbe etc. to get the values.
ProbeInterface is derived from BaseObject.

QuotientColProbe has been replaced by QuotientProbe, which is also
derived from ProbeInterface. It now takes any ProbeInterface-derived
objects for numerator and denominator, not just ColumnEnergyProbe
objects.

The probes' tasks are generally divided into three tasks, each
managed by their own class hierarchy.

The Local classes compute the local portion of the probe, without
performing any MPI. StatsProbe and the norm-base probes use this
component. Probes such as ColumnEnergyProbe and QuotientProbe do not
have a local component, because the input to these probes has already
been aggregated.

The Aggregator classes sum the local portions for each batch element.
StatsProbe uses this class either during the outputState stage or
during checkpointing, depending on a newly added boolean parameter,
immediateMPIAssembly. Setting this flag to false pushes the MPI needed
for the aggregation to the checkpointing, which is expected to improve
performance. Keeping it at false allows the many StatsProbe-derived
classes used in the system tests to continue to work with much less
refactoring. In fact, a new class, StatsProbeImmediate, was introduced
to keep automatically set this flag to true.

The ProbeInterface-derived classes perform the aggregation during
outputState, so that the results are available to probes that depend
on them, such as AdaptiveTimeScaleProbe.

Finally, the Outputter classes relay the aggregated probe values to the
MPI processes that to input/output for the IO MPIBlock. In an M-to-N MPI
configuration, the root process for the local communicator may not be
the root process for the IO block communicator. This task is performed
during checkpointing; the aggregator classes buffer the results of the
probe for each timestep until the checkpoint. This is expected to cut
down on the MPI traffic required by probes during large-scale M-to-N
runs.

As of this commit, the statsFlag has not been reimplemented. Hence the
ProbeStatsFlagTest system test fails. The statsFlag is used by
non-StatsProbe probe classes. If this boolean parameter is set, the
probe is supposed to output an average across all batch elements, as
opposed to outputting the values for each batch element individually.
This functionality will need to be restored in a future commit.

Many unit tests have been added for the new classes that form
components of the refactored probes. However, not all the classes have
unit tests. This should be remedied in a future commit.

This commit also removes deprecated probes from the list of files to be
compiled by CMake, and makes many other changes to update the
code to conform with preferred style.