Restores the old version of the tutorial lua file
The old version of the .lua file used in the LCA_CIFAR tutorial is now
at LCA_CIFAR_full. It uses nbatch = 16 and numImages = 50000,
to run for the entire CIFAR-10 dataset.

This way the old version of the tutorial on the wiki page is still
usable, while the new tutorial is being developed.