Removes long-obsolete dataType from params file
This removes the HyPerLayer dataType parameter from .params files,
which had remained even though this parameter has not been read or
used for several years.

It is not to be confused with the dataType field of .pvp file headers,
which is still used to identify whether a .pvp file is an activity
file, a sparse activity file, a shared weights file, or a nonshared
weights file.