broadcastFlag parameter added to HyPerLayer
If broadcastFlag for a layer is set to true, nxScale and nyScale are
not read. Instead, each process in the MPI configuration has local
nx and ny set to one, and the LayerInputBuffer synchronizes the
MPI processes using MPI_Allreduce and MPI_SUM. This layer is then
effectively a 1-by-1-by-nf layer.

A connection with sharedFlag=false and nxp=nyp=1 can therefore connect
a broadcastFlag=false HyPerLayer to a broadcastFlag=true layer as an
all-to-all connection.

If there are any layers with broadcastFlag=true, there are restrictions
on the possible sizes of the other layers: nx/numColumns and ny/numRows
must each be an integer power of two, where nx and ny are the HyPerCol
parameters for the base layer size, and numRows and numColumns are the
number of rows and columns in the MPI configuration.