CMake variable PV_DIR points to top level of repo
PV_DIR now points to the top level of the repository, instead of the
source directory. Generated .lua files use this directory to find
the parameterWrapper/?.lua files when the environmental variable
PV_SOURCEDIR is not present.

The internal CMake variables PV_SOURCE_DIR and PV_INCLUDE_DIR still
point to the src directory.