Fix weight normalization bug for all-zero patches
If NormalizeL2 encounters a weight patch with all weights equal to zero,
it should skip that weight patch and move onto the next patch. The
code was instead breaking out of the loop over patches, so that
subsequent patches would not get normalized, even though they should be.
This commit addresses that bug.