Change args for one of the Weights constructors
Weights now has a constructor that takes as arguments the name and an
existing Weights object. It uses the geometry (but not the weight
values) as initialization.

The Weights constructor that took only a name argument has been
removed.

The Weights::initialize() functions have been changed from public
to private; they are called by the constructor and do not have
to be called separately.

This simplifies the public interface for Weights.

Two system tests that used public Weights::initialize have been
refactored accordingly.