New layer type RotateLayer, and system test
RotateLayer computes the V buffer the same way HyPerLayer does, and
generates the activity buffer by rotating V.

The angle is generated randomly from a uniform distribution, with max
and min specified in params as angleMax and angleMin. The angleUnits
parameter specifies whether angleMax and angleMin are in degrees or
radians.