RotateLayer checkpoints the random state.
The root process of each IO MPI block gathers the RNGs of the
processes in its block, and writes them as a series of unsigned ints,
to the file "[Layer name]_rand_state.bin". A .txt file is also
created in human-readable format, but reading from checkpoint only
uses the .bin file.

There is a system test, RotateLayerRandStateTest, for this feature.