Change I/O for pre-synaptic broadcast connections
If a connection's pre-synaptic layer has its broadcastFlag set, the
files read/written by the connection are written as pre-synaptic layer
size of 1-by-1-by-nfPre, regardless of MPI configuration.

This is done by means of a new class BroadcastPreWeightsFile, and a
helper class, BroadcastPreWeightsFileIO.

There is a new enum class defined in the Weights class, WeightsType.
Its allowed values are
SHARED - used for shared weights
LOCALPATCH - used for local-patch weights, i.e. nonshared, nonbroadcast
BROADCASTPRE - used when the presynaptic layer is a broadcast layer

There are other minor edits and bugfixes, as well.