prevDelta.pvp can be written as broadcast-pre
If the presynaptic layer is a broadcast layer, MomentumConn will write
the .prevDelta.pvp file to the output directory as a broadcast-pre file.
Without this update, it would be written as a local-patch file, with
more than one patch in the x- and y- dimensions, corresponding to the
size of the MPI configuration.