Improvements to rescaling and interpolation
A warning is issued if the autoResizeFlag parameter is on in a context
where resizing is not necessary.

It is a fatal error if rescaling requires interpolating when either
the dimensions of the input image or the input layer are incompatible
with the interpolation method.

If using nearest-neighbor interpolation, the layer dimension (either
height or width) cannot be 1 unless the corresponding image dimension
is also 1.

For bicubic interpolation, the input height and width and the
layer height and width must all be at least 2 pixels.

Other minor changes have been made as well.