Use expm1() in HyPerLCALayer computation
Typically, the timeConstantTau parameter is much larger than the
adaptive time step, so that exp(-dt/tau) is close to one.  The presence
of 1-exp(-dt/tau) in the update rule for HyPerLCALayer therefore leads
to loss of precision if calculated using std::exp().

This update makes use of the standard library expm1() function for
computing exp(x)-1 without first computing exp(x).