Skeleton of InputVolumeLayer
This layer will take as input a list of directories, where each
directory contains a list of image files, and will interpret the
contents of one directory as a volume. With this commit, the
InputVolumeLayer compiles and can be added to a params file, but the
layer doesn't do anything yet.