Bugfix to BroadcastLayerFile::read()
There was an arithmetic error in handling the MPI batch index when
reading a broadcast layer file. The system tests did not handle a
case where this bug would arise. Several tests have been added
to the system tests, to verify that this bug did not exist in any other
I/O class.