Uniform random initialization defaults max to 1
This applies to both initializing a layer's V buffer and to
initializing weights. Previously, the default value of the max
was 1 plus the value of the min. Now, the default is taken from
the default params file, regardless of what the min value is.

DiscreteUniformRandom errors out if max <= min or if numValues is
less than 2.

UniformRandom errors out if max < min.