Change in handling of writeStep parameter
A writeStep of 0 now means do not write any output.
Previously, writeStep=0 had the effect of writing every timestep, and
a negative value meant not to write any output. Now a zero or
negative value means not to write any output. A value equal to the
HyPerCol's dt means to write every timestep (as it did before),
2*dt means every other timestep, and so forth.

The default is now taken from the default params file, if it exists.
(If no default params file is specified and writeStep is not specified
in the params file, it is a fatal error.)
Previously, writeStep was set equal to the HyPerCol's dt.