More efficiently constructs objects from params
Previously, the HyPerCol would create a ParamsIO object for
each parameter group, then extract the params and defaults from it
and pass it to the constructor, which would then take the params and
defaults and wrap them into a ParamsIO object. Now, the ParamsIO
object is passed directly to the constructor.