Change to PVParams::lookForUnread() function
lookForUnread() now returns a vector of pairs of strings, each pair
containing the name the param group and the name of the parameter for a
parameter that has not been read.

Previously, it returned 0 if all had been read or 1 if some had not been
read, and printed messages for any unread parameter to the error stream.

Now it returns the information about unread params; it is up to the
calling routine how to handle that information.

HyPerCol now handles the errors/warnings for unread params, and errors
out if errorOnUnusedParam is true, instead of PVParams doing so.