Relaxes size restrictions conns impose on layers
When one of the layers joined by a connection is a broadcast layer,
there are no restrictions on the dimensions of the other layer.

Previously, the requirement that sizes of nonbroadcast layer sizes be
related by a power of two was applied in the broadcast case as well,
even though there is no conceptual issue with connecting arbitrary
sizes to a 1-by-1 layer.