New error condition for group normalization
Connections with normalizeMethod = normalizeGroup
do not use any of the parameters
   normalizeArborsIndividually
   normalizeOnInitialize
   normalizeOnWeightUpdate
   strength
Instead, the group head applies its parameters to all connections
in the group, and the other connections in the group ignore these
parameters.

With this update, it is a fatal error for connections in a
normalization group to have different values of the above parameter.

One potential issue is that if the group head sets strength to a
nondefault value and the other connections in the group do not
set the strength value, the generated params file lists the strength
for the other connections as having the default value. Running
from the generated params file without modification will therefore
trip an error.

This issue is difficult to resolve because strength is read by the
Gauss2DWeight initialization method as well as by normalizations. For
this reason the strength parameter is not directly read by the
normalization class and it is complicated to determine whether the
strength parameter should be included or not.